refactor: subcommands, handler and ipc codes are refactored into sub folders
