+++ src/daemon/lifecycle/daemonManager.ts
