+++ src/daemon/lifecycle/serviceInstaller.ts
