+++ src/daemon/lifecycle/sessionTracker.ts
