+++ src/daemon/lifecycle/stateStore.ts
